+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Auth commands are now available via `plistsync auth [service]` instead of `plistsync [service] auth`
 - Added `plistsync --version` command to show the currently installed version of the library
+- Renamed `_apply_diff` -> `_remote_commit` for clarity
+- Split Playlist ABC into two classes: one for simple services, like filesystems, where states can be pushed via a single API call (`PlaylistCollection`) and one where multiple API calls are required (`MultiRequestPlaylistCollection`), e.g. when a playlists description cannot be pushed in the same call as track changes.
 
 ## [0.5.1] - 2026-03-16